fix: reuse stable scrollbar for planting plans
